requirements.txt
altgraph==0.17.4
babel==2.17.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pywin32-ctypes==0.2.3
